feat: grid filtering and oredering + lot of little addons like kill + some fixes (2)

+++ packages/backend/src/main.rs
@@ -47,7 +47,7 @@ async fn run(gateway: Arc<Mutex<Gateway>>) {
         let current_time = std::time::SystemTime::now()
             .duration_since(std::time::UNIX_EPOCH)
             .unwrap()
-            .as_millis() as u64;
+            .as_millis();
         // ===== Heartbeat =====
         let mut remove_idxs: Vec<usize> = Vec::new();
         println!(
